refactor: extract functions and add coverage
